Introduce macrobatches for more verbose output
Motivation: I want to get asynchronous data iteration while not blocking
on net.fit() until the end of epoch. This way the learn process is
asynchronous inside one macrobatch and synchronous across all the
macrobatches.

Macrobatch is a group of minibatches. The iterator is modified so that
it reports the end of data when it exhausts a macrobatch. Then it
advances (manually) to the next macrobatch.